initial commit: html redirect & readme
The old repository moved to
https://github.com/near/rainbow-bridge-frontend

This set up a GitHub redirect when people visit
https://github.com/near-examples/erc20-to-nep21, but when people visit
the old website at https://near-examples.github.io/erc20-to-nep21, it
did not set up a corresponding redirect

This repository creates such a redirect, with the trade-off that
directly visiting the new repository link will show a README instead
of automatically redirecting. This seems like a worthwhile tradeoff.